Fix Codefactor issues
