+++ Quaver.Shared/Screens/Edit/UI/Footer/Bookmarks/EditorFooterBookmarkContainer.cs
@@ -1,33 +1,24 @@
-using System;
 using System.Collections.Generic;
-using Microsoft.Xna.Framework;
 using Quaver.API.Maps.Structures;
-using Quaver.Shared.Assets;
-using Quaver.Shared.Graphics;
-using Quaver.Shared.Screens.Edit.Actions.Bookmarks;
 using Quaver.Shared.Screens.Edit.Actions.Bookmarks.Add;
 using Quaver.Shared.Screens.Edit.Actions.Bookmarks.Edit;
 using Quaver.Shared.Screens.Edit.Actions.Bookmarks.Offset;
 using Quaver.Shared.Screens.Edit.Actions.Bookmarks.Remove;
-using Quaver.Shared.Screens.Edit.Dialogs;
-using Wobble;
 using Wobble.Graphics;
-using Wobble.Graphics.UI.Buttons;
-using Wobble.Graphics.UI.Dialogs;
 using Wobble.Window;
 
-namespace Quaver.Shared.Screens.Edit.UI.Footer
+namespace Quaver.Shared.Screens.Edit.UI.Footer.Bookmarks
 {
     /// <summary>
     ///     Container for displaying bookmarks across the width of the screen
     /// </summary>
-    public class EditorFooterBookmarks : Container
+    public class EditorFooterBookmarkContainer : Container
     {
         private EditScreen Screen { get; }
 
         private Dictionary<BookmarkInfo, EditorFooterBookmark> Bookmarks { get; } = new Dictionary<BookmarkInfo, EditorFooterBookmark>();
 
-        public EditorFooterBookmarks(EditScreen screen)
+        public EditorFooterBookmarkContainer(EditScreen screen)
         {
             Screen = screen;
             Size = new ScalableVector2(WindowManager.Width, 14);
@@ -93,43 +84,6 @@ namespace Quaver.Shared.Screens.Edit.UI.Footer
                 RemoveBookmark(bookmark);
                 AddBookmark(bookmark);
             }
-        }
-    }
-
-    /// <summary>
-    ///     An individual bookmark display object
-    /// </summary>
-    public class EditorFooterBookmark : ImageButton
-    {
-        private EditScreen Screen { get; }
-        private BookmarkInfo Bookmark { get; }
-
-        private Tooltip Tooltip { get; }
-
-        public EditorFooterBookmark(EditScreen screen, BookmarkInfo bookmark) : base(UserInterface.BlankBox)
-        {
-            Screen = screen;
-            Bookmark = bookmark;
-            Tint = Color.Yellow;
-            Tooltip = new Tooltip(Bookmark.Note, Tint) { DestroyIfParentIsNull = false };
-
-            Hovered += (sender, args) =>
-            {
-                if (string.IsNullOrEmpty(Bookmark.Note))
-                    return;
-                
-                screen.ActivateTooltip(Tooltip);
-            };
-            
-            LeftHover += (sender, args) => screen.DeactivateTooltip();
-            Clicked += (sender, args) => DialogManager.Show(new EditorBookmarkDialog(Screen.ActionManager, Screen.Track, Bookmark)); 
-            RightClicked += (sender, args) => screen.ActionManager.RemoveBookmark(Bookmark);
-        }
-
-        public override void Destroy()
-        {
-            Tooltip.Destroy();
-            base.Destroy();
         }
     }
 }
